renamed panels folder
