+++ swgs/panels/panelapp_panels.py
